+++ lib/states/state_manager.dart
@@ -2,6 +2,7 @@ import 'package:defend_your_flame/constants/translations/app_strings.dart';
 import 'package:defend_your_flame/core/flame/game_provider.dart';
 import 'package:defend_your_flame/helpers/platform_helper.dart';
 import 'package:defend_your_flame/widgets/background.dart';
+import 'package:defend_your_flame/widgets/platform_wrappers/web_wrapper.dart';
 import 'package:flame/game.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
@@ -29,45 +30,19 @@ class _StateManagerState extends State<StateManager> with WidgetsBindingObserver
           '/': (context) {
             var game = GameProvider.of(context).game;
             game.setExternalDependencies(context);
-            return Scaffold(
-                body: Center(
-                    child: PopScope(
-                        canPop: false,
-                        child: _buildPlatformDimensionsIfNeeded(
-                            game: Padding(
-                          padding: EdgeInsets.all(PlatformHelper.borderPadding),
-                          child: GameWidget(
-                            game: game,
-                            backgroundBuilder: (context) => const Background(),
-                          ),
-                        )))));
+
+            Widget gameWidget = GameWidget(
+              game: game,
+              backgroundBuilder: (context) => const Background(),
+            );
+
+            if (PlatformHelper.isWeb) {
+              gameWidget = WebWrapper(child: gameWidget);
+            }
+
+            return Scaffold(body: PopScope(canPop: false, child: gameWidget));
           }
         });
-  }
-
-  Widget _buildPlatformDimensionsIfNeeded({
-    required Widget game,
-  }) {
-    var maxWidth = PlatformHelper.maxRenderWidth;
-    var maxHeight = PlatformHelper.maxRenderHeight;
-
-    if (maxWidth == null || maxHeight == null) {
-      return game;
-    }
-
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        Expanded(
-            child: ClipRect(
-          child: Container(
-            constraints: BoxConstraints(maxWidth: maxWidth, maxHeight: maxHeight),
-            child: game,
-          ),
-        )),
-        if (PlatformHelper.isWeb) PlatformHelper.webRedirectFooter()
-      ],
-    );
   }
 
   @override